Module Disk_persistence.Make_heap

Create an on-disk persistent version of the Bounded_min_heap data structure.

Parameters

module N : sig ... end

module K : Tezos_crypto.Intfs.HASH

module V : sig ... end

Signature

type t

val remove : t -> K.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

remove h k removes the binding from k in h. If k is not bound in c, it does nothing. The removal is persisted on disk.

val remove_predicate : (V.t -> bool) -> t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

remove_predicate f h removes the binding in h where f elt = true. The removal is persisted on disk.

val insert : t -> K.t -> V.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

insert h K.t v binds the key k to the value v in the queue h. This may or may not cause another binding to be removed, depending on the number of bindings already present in h. The addition (or replacement) is persisted on disk.

val peek_min : t -> V.t option

peek_min h return, without removing, the smallest element from the heap h, None if empty.

val pop : t -> V.t option Tezos_base.TzPervasives.tzresult Lwt.t

pop h remove the smallest element from the heap h, None if empty.

val find_opt : t -> K.t -> V.t option

find_opt h k is Some v if k is bound to v in h. It is None otherwise.

val elements : t -> V.t list

elemets h returns the elements of the queue h from oldest to newest.

val length : t -> int

length h is the number of bindings held by h.

val clear : t -> unit Lwt.t

clear h empties the queue h and removes its persistent content on disk.

val load_from_disk : warn_unreadable: (string -> Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace -> unit Lwt.t) option -> capacity:int -> data_dir:string -> filter:(V.t -> bool) -> t Tezos_base.TzPervasives.tzresult Lwt.t

load_from_disk ~warn_unreadable ~capacity ~data_dir ~filter creates a bounded hash queue of capacity capacity. The queue is populated by persistent elements present in data_dir/N.name which pass the filter (the directory is created if it does not exist). If warn_unreadable is Some warn, unreadable files are ignored but a warning is printed with warn, otherwise the loading fails on the first unreadable file.