Module Octez_injector.Injector_common

type signer = {

alias : string;

pkh : Tezos_base.TzPervasives.Signature.public_key_hash;

pk : Tezos_base.TzPervasives.Signature.public_key;

sk : Tezos_client_base.Client_keys.sk_uri;

}

The type of signers for operations injected by the injector

val get_signer : Tezos_client_base.Client_context.wallet -> Tezos_base.TzPervasives.Signature.public_key_hash -> signer Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve a signer from the client wallet.

type tez = {

mutez : int64;

}

tez representation

type fee_parameter = {

minimal_fees : tez;

minimal_nanotez_per_byte : Q.t;

minimal_nanotez_per_gas_unit : Q.t;

force_low_fee : bool;

fee_cap : tez;

burn_cap : tez;

}

fee parameters for each operation type

val fee_parameter_encoding : default_fee_parameter:fee_parameter -> fee_parameter Tezos_base.TzPervasives.Data_encoding.t

fee_parameter_encoding default_parameter_encoding generates the encoding using default value from default_fee_parameter