Module Octez_injector.Injector_errors

type Tezos_base.TzPervasives.error +=

| No_worker_for_source of Tezos_base.TzPervasives.Signature.Public_key_hash.t

Error when the injector has no worker for the source which must inject an operation.

type Tezos_base.TzPervasives.error +=

| No_worker_for_tag of string

Error when the injector has no worker for the tag of the operation to be injected.

type Tezos_base.TzPervasives.error +=

| Step_failed of string

Error when a step of the injector failed.

type Tezos_base.TzPervasives.error +=

| Overlapping_operations

Error when multiple workers are assigned to the same operation.