Module P.Operation

A module for the injector operations

type t

The abstract type of operations to inject

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

An encoding for injector's operations

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printing injector's operations

val unique : t -> bool

If unique op = true, then the injector's queue will only contain at most one operation op. Otherwise, it will allow duplicate such operations to appear in the queue.

val compare : t -> t -> int