Module type Injector_sigs.PARAMETERS

Module type for parameter of functor Injector_functor.Make.

type state

The type of the state that the injector can access

module Tag : TAG

A module which contains the different tags for the injector

module Operation : PARAM_OPERATION

A module for the injector operations

val events_section : string list

Where to put the events for this injector

val table_estimated_size : Tag.t -> int

Coarse approximation for the number of operation of each tag we expect to inject for each block.

val retry_unsuccessful_operation : state -> Operation.t -> ?reason: (Operation.t * Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) -> unsuccessful_status -> retry_action Lwt.t

Action (see retry_action) to be taken on unsuccessful operation (see unsuccessful_status).

val operation_tag : Operation.t -> Tag.t

The tag of a manager operation. This is used to send operations to the correct queue automatically (when signer is not provided) and to recover persistent information.

val fee_parameter : state -> Operation.t -> Injector_common.fee_parameter

Returns the fee_parameter (to compute fee w.r.t. gas, size, etc.) and the caps of fee and burn for each operation.

val safety_guard : Operation.t -> int option

Returns the gas safety guard for each operation if it should be different from the client default (100).

val persist_operation : Operation.t -> bool

Indicate which operations should be persisted on disk to be reinjected upon restart.

val metrics_registry : Prometheus.CollectorRegistry.t