Module type Injector_sigs.PROTOCOL_CLIENT

type state

type operation

type unsigned_operation

val max_operation_data_length : int

val manager_pass : int

The validation pass of manager operations.

val operation_status : state -> Tezos_base.TzPervasives.Block_hash.t -> Tezos_base.TzPervasives.Operation_hash.t -> index:int -> operation_status option Tezos_base.TzPervasives.tzresult Lwt.t

operation_status block oph ~index returns the status of the operation at position index in the L1 batch oph included in the block block. It returns None if the operation with the given index is not in the block.

val operation_size : operation -> int

Size of an operation in bytes according to the protocol. This only accounts for the actual content of the corresponding manager operation (and not its fees, gas, etc.).

val operation_size_overhead : int

An upper bound of the overhead added to manager operations in bytes. Typically, this would include the source, fees, counter, gas limit, and storage limit.

val simulate_operations : Tezos_client_base.Client_context.full -> force:bool -> source:Tezos_base.TzPervasives.Signature.public_key_hash -> src_pk:Tezos_base.TzPervasives.Signature.public_key -> successor_level:bool -> fee_parameter:Injector_common.fee_parameter -> ?safety_guard:int -> operation list -> (unsigned_operation simulation_result, [ `Exceeds_quotas of Tezos_base.TzPervasives.tztrace | `TzError of Tezos_base.TzPervasives.tztrace ]) Stdlib.result Lwt.t

Simulating a batch of operations. This function returns the simulation result for each of these operations (with its associated index in the batch, in case there is a revelation operation added) together with a Tezos raw unsigned operation that can be directly injected on a node if one wishes to.

val sign_operation : Tezos_client_base.Client_context.full -> Tezos_client_base.Client_keys.sk_uri -> unsigned_operation -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

Sign an unsigned operation an return the serialized signed operation, ready for injection.

val time_until_next_block : state -> Tezos_base.Block_header.shell_header option -> Ptime.span

time_until_next_block state block_header computes the time until the block following block_header, with respect to the current time.

val checks : state -> unit Tezos_base.TzPervasives.tzresult

Run protocol specific checks for injector configuration/state.

val get_balance_mutez : Tezos_client_base.Client_context.full -> ?block:Tezos_base.TzPervasives.Block_hash.t -> Tezos_base.TzPervasives.Signature.public_key_hash -> int64 Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the balance in mutez for a given public key hash.