Module S.Inj_operation

type operation = operation

type id = Id.t

Alias for L1 operations ids

type errors = {

count : int;

last_error : Tezos_base.TzPervasives.tztrace option;

}

Structure to keep track of injection errors.

type t = private {

id : id;

order : Z.t option;

counter : Z.t;

operation : operation;

mutable errors : errors;

}

The type of L1 operations that are injected on Tezos. These have an id attached to them that allows tracking and retrieving their status.

val make : ?order:Z.t -> operation -> t

make op returns an L1 operation with the corresponding hash.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

Encoding for L1 operations

val pp : Stdlib.Format.formatter -> t -> unit

Pretty printer for L1 operations. Only the relevant part for the rollup node is printed.

val register_error : t -> Tezos_base.TzPervasives.tztrace -> unit

Register an error as occurring during injection of an operation. Its internal error counter is incremented.

val id : t -> id

val compare : t -> t -> int