Module Injector_worker_types.Request

Parameters

module Tag : Injector_sigs.TAG

module L1_operation : Injector_sigs.INJECTOR_OPERATION

Signature

type removal_criteria =

| Operation_tag of Tag.t
(*
Operation_tag kind Clear all operation of type kind
*)

| Order_below of Z.t * bool
(*
Order_below (order, drop_no_order Clear all operation that have an order specified below order. If drop_no_order also clear all operation that has no order specified.
*)

| Tag_and_order_below of Tag.t * Z.t * bool
(*
Tag_and_order_below (kind, order, drop_no_order) Apply Order_below but only for operation of type kind.
*)

Criteria to specify which operation to remove from the injector.

type ('a, 'b) t =

| Inject : (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t

| Clear : removal_criteria option -> (unit, Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace) t

type view =

| View : (_, _) t -> view

include Tezos_base.Worker_intf.REQUEST with type ('a, 'request_error) t := ('a, 'request_error) t and type view := view

include Tezos_base.Worker_intf.VIEW with type view := view

val encoding : view Data_encoding.t

val pp : Stdlib.Format.formatter -> view -> unit

val view : ('response, 'error) t -> view

The projection function from full request to simple views.