Module Metrics.Make

Parameters

module P : P

Signature

val produce_metrics : bool -> unit

Set a flag that will let executing the code to produce metrics through the wrap function

val wrap : (unit -> unit) -> unit

Wrap metrics code to be executed only if the metrics server is on

val add_gauge : int -> P.Tag.t list -> unit

Add queue, injected and included gauge for a tag

val set_queue_size : P.Tag.t list -> int -> unit

Set the operations queue size

val set_injected_operations_size : P.Tag.t list -> int -> unit

Set the injected operations size

val set_included_operations_size : P.Tag.t list -> int -> unit

Set the included operations size